Centralize _normalize_binop_value.
